Simplifying metric goal category aggregation fix: removed SyntheticFeature goals fix: removed InputQuality goals fix: updated metric goal cateogries in all metrics fix: wrong data casting in image feature extractor
