--- requirements/requirements-text.txt
@@ -1,6 +1,0 @@
-# accelerate>0.24.1, <0.25.0 ; python_version >= "3.9" and python_version < "3.13"
-# datasets==2.20.0 ; python_version >= "3.9" and python_version < "3.13"
-# nltk==3.8.1 ; python_version >= "3.9" and python_version < "3.13"
-# python-multipart==0.0.6 ; python_version >= "3.9" and python_version < "3.13"
-# sentence-transformers==2.7.0 ; python_version >= "3.9" and python_version < "3.13"
-# transformers==4.43.2 ; python_version >= "3.9" and python_version < "3.13"
