+++ scripts/tabular_metrics.sh
@@ -7,14 +7,12 @@ pymdma --modality tabular \
     --target_data data/test/tabular/input_val/dataset \
     --batch_size 1\
     --output_dir reports/tabular_metrics/ \
-    --metric_group quality privacy
 
-# pymdma --modality tabular \
-#     --validation_type synth \
-#     --evaluation_level dataset \
-#     --reference_type dataset \
-#     --reference_data data/test/tabular/synthesis_val/reference \
-#     --target_data data/test/tabular/input_val/dataset \
-#     --batch_size 1\
-#     --output_dir reports/tabular_metrics/ \
-#     --metric_group feature data
+pymdma --modality tabular \
+    --validation_type synth \
+    --evaluation_level dataset \
+    --reference_type dataset \
+    --reference_data data/test/tabular/synthesis_val/reference \
+    --target_data data/test/tabular/input_val/dataset \
+    --batch_size 1\
+    --output_dir reports/tabular_metrics/ \
