+++ src/pymdma/common/compute.py
@@ -48,6 +48,7 @@ class ComputationManager:
         self.n_workers = n_workers
 
         self.metrics = self._instanciate_metric_classes(group_classes)
+
         self.global_context: Dict[str, any] = {}
 
         self.extractors = set() if pretrained_extractor_name is None else {pretrained_extractor_name}
@@ -109,7 +110,7 @@ class ComputationManager:
 
         def _compute_task(metric, metric_args):
             metric_name = metric.__class__.__name__
-            logger.info(f"Extractor: {metric.extractor_model_name} | Metric: {metric_name}")
+            logger.info(f"Metric: {metric_name}")
             new_result = metric.compute(*metric_args, context=self.global_context)
 
             # merge metric with already compute one (batch calculation)
