+++ src/pymdma/constants.py
@@ -1,5 +1,4 @@
 from enum import Enum
-from typing import Union
 
 METRICS_PACKAGE_NAME = "measures"
 SEED = 42
@@ -29,6 +28,13 @@ class ValidationTypes(str, StrEnum):
     SYNTH = "synthesis_val"
 
 
+class MetricGoal(str, StrEnum):
+    VALIDITY = "validity"
+    QUALITY = "quality"
+    PRIVACY = "privacy"
+    UTILITY = "utility"
+
+
 ################################################################
 ####################### SYNTHETIC GROUP ########################
 ################################################################
@@ -37,52 +43,12 @@ class SyntheticMetricGroups(str, StrEnum):
     DATA = "data"
 
 
-# ============ Synthetic Metric Specific Goals ============
-class SyntheticFeatureMetrics(str, StrEnum):
-    FIDELITY = "fidelity"
-    DIVERSITY = "diversity"
-    AUTHENTICITY = "authenticity"  # indicates privacy
-    PRIVACY = "privacy"
-    UTILITY = "utility"  # TODO move utility to another group (not based on features)
-    QUALITY = "quality"  # general quality
-
-
 ################################################################
 ####################### INPUT GROUP ############################
 ################################################################
 class InputMetricGroups(str, StrEnum):
-    QUALITY = "quality"
+    QUALITY = "data"
     ANNOTATION = "annotation"
-    PRIVACY = "privacy"
-
-
-# ============ Input Metric Specific Goals ===============
-class InputQualityMetrics(str, StrEnum):
-    CONTRAST = "contrast"
-    BRIGHTNESS = "brightness"
-    COLORFULNESS = "colorfulness"
-    SHARPNESS = "sharpness"
-    PERCEPTUAL_QUALITY = "perceptual_quality"
-    NOISE = "noise"
-    SIMILARITY = "similarity"
-    UNIFORMITY = "uniformity"
-    UNIQUENESS = "uniqueness"
-    CONSISTENCY = "consistency"
-    DIMENSIONALITY = "dimensionality"
-
-
-class InputPrivacyMetrics(str, StrEnum):
-    PRIVACY = "privacy"
-    ANONYMITY = "anonymity"
-    CONFIDENTIALITY = "confidentiality"
-    NON_REPUDIATION = "non_repudiation"
-    UNIQUENESS = "uniqueness"
-
-
-class InputAnnotationMetrics(str, StrEnum):
-    COMPLETENESS = "completeness"
-    CORRECTNESS = "correctness"
-    UNIQUENESS = "uniqueness"
 
 
 # ============ Annotation Types ===============
@@ -93,27 +59,15 @@ class AnnotationType(str, StrEnum):
     KEYPOINTS = "keypoints"
 
 
-def valid_subclass(goal, subgoal):
-    dependency_map = {
-        InputMetricGroups.QUALITY: InputQualityMetrics,
-        InputMetricGroups.ANNOTATION: InputAnnotationMetrics,
-        InputMetricGroups.PRIVACY: InputPrivacyMetrics,
-        SyntheticMetricGroups.FEATURE: SyntheticFeatureMetrics,
-    }
-    return goal in dependency_map and dependency_map[goal].has_value(subgoal)
-
-
 # ===================== OTHER CONFIGS =====================
 class EvaluationLevel(str, StrEnum):
     DATASET = "dataset"
     INSTANCE = "instance"
-    # FEATURE = "feature_wise"
 
 
 class ReferenceType(str, StrEnum):
     DATASET = "dataset"
     INSTANCE = "instance"
-    # FEATURE = "feature"
     NONE = "none"
 
 
@@ -125,6 +79,3 @@ class OutputsTypes(str, StrEnum):
     STRING = "string"
     KEY_VAL = "key_value"  # key-value pair (dict) str -> float | int | str
     KEY_ARRAY = "key_array"  # key-values pair (dict) str -> list of float | int | str
-
-
-METRIC_GOALS = Union[InputQualityMetrics, InputPrivacyMetrics, InputAnnotationMetrics, SyntheticFeatureMetrics]
