+++ src/pymdma/general/measures/prdc.py
@@ -2,7 +2,7 @@ import numpy as np
 
 from pymdma.common.definitions import FeatureMetric
 from pymdma.common.output import MetricResult
-from pymdma.constants import EvaluationLevel, OutputsTypes, ReferenceType, SyntheticFeatureMetrics
+from pymdma.constants import EvaluationLevel, MetricGoal, OutputsTypes, ReferenceType
 
 from ..utils.util import compute_nearest_neighbour_distances, compute_pairwise_distance
 
@@ -43,7 +43,7 @@ class ImprovedPrecision(FeatureMetric):
 
     reference_type = ReferenceType.DATASET
     evaluation_level = [EvaluationLevel.INSTANCE, EvaluationLevel.DATASET]
-    metric_goal = SyntheticFeatureMetrics.FIDELITY
+    metric_goal = MetricGoal.QUALITY
 
     higher_is_better: bool = True
     min_value: float = 0.0
@@ -136,7 +136,7 @@ class ImprovedRecall(FeatureMetric):
 
     reference_type = ReferenceType.DATASET
     evaluation_level = [EvaluationLevel.INSTANCE, EvaluationLevel.DATASET]
-    metric_goal = SyntheticFeatureMetrics.DIVERSITY
+    metric_goal = MetricGoal.QUALITY
 
     higher_is_better: bool = True
     min_value: float = 0.0
@@ -230,7 +230,7 @@ class Density(FeatureMetric):
 
     reference_type = ReferenceType.DATASET
     evaluation_level = [EvaluationLevel.INSTANCE, EvaluationLevel.DATASET]
-    metric_goal = SyntheticFeatureMetrics.FIDELITY
+    metric_goal = MetricGoal.QUALITY
 
     higher_is_better: bool = True
     min_value: float = 0.0
@@ -320,7 +320,7 @@ class Coverage(FeatureMetric):
 
     reference_type = ReferenceType.DATASET
     evaluation_level = [EvaluationLevel.INSTANCE, EvaluationLevel.DATASET]
-    metric_goal = SyntheticFeatureMetrics.DIVERSITY
+    metric_goal = MetricGoal.QUALITY
 
     higher_is_better: bool = True
     min_value: float = 0.0
@@ -421,7 +421,7 @@ class Authenticity(FeatureMetric):
 
     reference_type = ReferenceType.DATASET
     evaluation_level = [EvaluationLevel.INSTANCE, EvaluationLevel.DATASET]
-    metric_goal = SyntheticFeatureMetrics.AUTHENTICITY
+    metric_goal = MetricGoal.PRIVACY
 
     higher_is_better: bool = True
     min_value: float = 0.0
