+++ src/pymdma/image/measures/input_val/data/reference.py
@@ -8,7 +8,7 @@ from torchmetrics.image import MultiScaleStructuralSimilarityIndexMeasure, Struc
 
 from pymdma.common.definitions import Metric
 from pymdma.common.output import MetricResult
-from pymdma.constants import EvaluationLevel, InputQualityMetrics, OutputsTypes, ReferenceType
+from pymdma.constants import EvaluationLevel, MetricGoal, OutputsTypes, ReferenceType
 
 # TODO review documentations and attributes
 
@@ -36,7 +36,7 @@ class PSNR(Metric):
 
     reference_type = ReferenceType.INSTANCE
     evaluation_level = EvaluationLevel.INSTANCE
-    metric_goal = InputQualityMetrics.NOISE
+    metric_goal = MetricGoal.QUALITY
 
     higher_is_better: bool = True
     min_value: float = 0.0
@@ -122,7 +122,7 @@ class SSIM(Metric):
 
     reference_type = ReferenceType.INSTANCE
     evaluation_level = EvaluationLevel.INSTANCE
-    metric_goal = InputQualityMetrics.SIMILARITY
+    metric_goal = MetricGoal.QUALITY
 
     higher_is_better: bool = True
     min_value: float = 0.0
@@ -217,7 +217,7 @@ class MSSIM(Metric):
 
     reference_type = ReferenceType.INSTANCE
     evaluation_level = EvaluationLevel.INSTANCE
-    metric_goal = InputQualityMetrics.SIMILARITY
+    metric_goal = MetricGoal.QUALITY
 
     higher_is_better: bool = True
     min_value: float = 0.0
