+++ src/pymdma/image/models/extractor.py
@@ -30,12 +30,15 @@ class StandardTransform:
         image = self.preprocess_transform(image) if self.preprocess_transform is not None else image
         image = image.resize(self.img_size, self.interp)
         # bring image to the range [0, 1] and normalize to [-1, 1]
-        image = np.array(image).astype(np.float32) / 255.0
-        image = image * 2.0 - 1.0
-        return torch.from_numpy(image).permute(2, 0, 1)
+        image = (np.array(image) - 128) / 128
+        # image = np.array(image).astype(np.float32) / 255.0
+        # image = image * 2.0 - 1.0
+        return torch.from_numpy(image).permute(2, 0, 1).to(torch.float32)
 
 
 class BaseExtractor(torch.nn.Module, EmbedderInterface):
+    extractor: torch.nn.Module
+
     def __init__(
         self,
         input_size: Tuple[int],
@@ -68,7 +71,7 @@ class BaseExtractor(torch.nn.Module, EmbedderInterface):
             batch_sizes.append(len(files) % batch_size)
 
         transform = StandardTransform(self.input_size, self.interpolation)
-        self.extractor = self.extractor.to(device)
+        self.extractor = self.extractor.to(device, dtype=torch.float32)
 
         act_array = []
         start, end = 0, 0
@@ -101,10 +104,10 @@ class BaseExtractor(torch.nn.Module, EmbedderInterface):
         labels_array = []
         ids_array = []
 
-        self.extractor = self.extractor.to(device)
+        self.extractor = self.extractor.to(device, dtype=torch.float32)
         dataloader.dataset.transform = StandardTransform(self.input_size, self.interpolation, preprocess_transform)
         for batch, labels, img_ids in tqdm(dataloader, total=len(dataloader)):
-            batch = batch.to(device)
+            batch = batch.to(device, dtype=torch.float32)
             batch = self.extractor(batch).detach().cpu().numpy()
             act_array.append(batch)
             labels_array.append(labels)
