+++ src/pymdma/time_series/measures/input_val/__init__.py
@@ -1,3 +1,3 @@
-from pymdma.time_series.measures.input_val.quality import SNR, Uniqueness
+from pymdma.time_series.measures.input_val.data.quality import SNR, Uniqueness
 
 __all__ = ["SNR", "Uniqueness"]
