+++ src/pymdma/time_series/measures/input_val/data/quality.py
@@ -4,7 +4,7 @@ import numpy as np
 
 from pymdma.common.definitions import Metric
 from pymdma.common.output import MetricResult
-from pymdma.constants import EvaluationLevel, InputQualityMetrics, OutputsTypes, ReferenceType
+from pymdma.constants import EvaluationLevel, MetricGoal, OutputsTypes, ReferenceType
 
 
 class Uniqueness(Metric):
@@ -29,7 +29,7 @@ class Uniqueness(Metric):
 
     reference_type = ReferenceType.NONE
     evaluation_level = EvaluationLevel.INSTANCE
-    metric_goal = InputQualityMetrics.UNIQUENESS
+    metric_goal = MetricGoal.QUALITY
 
     higher_is_better: bool = False
     min_value: float = 0.0
@@ -89,7 +89,7 @@ class SNR(Metric):
 
     reference_type = ReferenceType.NONE
     evaluation_level = EvaluationLevel.INSTANCE
-    metric_goal = InputQualityMetrics.NOISE
+    metric_goal = MetricGoal.QUALITY
 
     higher_is_better: bool = False
     min_value: float = 0.0
